Wave 11: 128 functions matched, 885 -> 1013, and families are now computable
Seven agents on families found by scanning for repeated address spacing, one
agent building the tool that makes that scan unnecessary.

All 128 verified with trymatch before promotion; both builds still reproduce
SHA1 14dd0b22b894865867aff89e8116b2dffae25605.

The largest batch was 41 contiguous forwarders at 0x08071F88, which turned out
to be four templates. Their constants are a fade SPEED, not the bitmask the
brief guessed: arg 1 lands at proc+0x54, an accumulator at +0x5c adds it per
frame and terminates at 0x200, so 4/8/0x10/0x20/0x40 are 128/64/32/16/8-frame
fades.

tools/families.py clusters every function by normalised instruction shape.
Registers are kept in the key and callee names are erased -- both measured, not
assumed. Inside one ROM built once, agbcc's allocation is a function of the
source, so a register difference IS a source difference: on the 41-member
forwarder shape the register split is 34/3/3/1 and each shard is a different C
template, because the register holding the immediate names the argument
position and `pop {r0}` vs `pop {r1}` separates void from value-returning.
Erasing callees goes the other way -- keeping them shatters the family into
singletons. It rediscovers all three hand-found families and exits non-zero if
it ever stops doing so.

The result that changes wave 12: family F000 is 72 members, 71 unmatched,
spanning 0x08004B2C-0x080879A0, with the shape already matched in
c_080733B8.c. Address-locality scanning could not have found it.

Four claims in my own brief were wrong and the agents corrected them:

- sub_0802C62A is .align padding, not an odd-address entry point. The tell
  itself survives -- 3 of the ROM's 4 non-word-aligned entries are genuine --
  but I cited the one counter-example. Chasing it found sub_0806F856 and
  sub_0806FC0A unrecorded; both are now in asm-resident.json, so they stop
  being offered as free 12-byte leaves.
- The five 44-byte bitfield siblings do not vary in shift or compare at all.
  The discriminator is pointer spelling: p[X+C] and *(p+X+C) are different code
  when a pointer global's deref is used twice in one address expression.
- unknown-functions.h is not "ground truth". sub_0803B4DC and sub_0803B524
  were declared void f(s16) and are int. An s16 parameter folds the caller's
  ldrh into ldrsh -- a 2-byte change with no separate shift to spot. Corollary:
  a prologue narrowing is not a parameter-width tell when the narrowed value
  feeds another call.
- Argument setup at a call is grouped by operand class, not argument order:
  parameter copies, then pool ldrs, then mov #imm8. That 4-byte size signal is
  what splits the forwarder family into 20- and 24-byte halves.

Retyping four callees to u8 forced edits to already-promoted c_0803CA9C.c,
c_0803CAD4.c and c_0803CAF0.c. All still match: the values are ldrb-derived, so
gcc drops the return narrowing.

+++ src/decomp/c_0803CA9C.c
@@ -7,7 +7,7 @@
  * sub_0803CA9C @ 0x0803CA9C, sub_0803CAB8 @ 0x0803CAB8
  */
 
-int sub_0803CA9C(u32 id)
+u8 sub_0803CA9C(u32 id)
 {
     struct Unk02028030 *s = &gUnknown_02028030;
     u32 idx = id >> 3;
@@ -17,7 +17,7 @@ int sub_0803CA9C(u32 id)
     return (1 << (id & 7)) & *p;
 }
 
-int sub_0803CAB8(u32 id)
+u8 sub_0803CAB8(u32 id)
 {
     struct Unk02028030 *s = &gUnknown_02028030;
     u32 idx = id >> 3;
