+++ src/decomp/c_0803CAD4.c
@@ -14,7 +14,7 @@
  * sub_0803CAD4 @ 0x0803CAD4
  */
 
-int sub_0803CAD4(u32 id)
+u8 sub_0803CAD4(u32 id)
 {
     struct Unk02028030 *s = &gUnknown_02028030;
     u32 idx = id >> 3;
